requirements.txt
absl-py==2.1.0
aenum==3.1.15
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aioice==0.9.0
aiortc==1.10.1
aiosignal==1.3.1
altair==5.5.0
annotated-types==0.7.0
anyio==4.6.2.post1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
array-record==0.4.1
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
async-timeout==4.0.3
attrs==24.2.0
av==13.1.0
babel==2.16.0
backoff==2.2.1
bcrypt==4.2.1
beautifulsoup4==4.12.3
bing-image-downloader==1.1.2
bleach==6.2.0
blinker==1.8.2
branca==0.8.0
build==1.2.2.post1
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
chroma-hnswlib==0.7.6
chromadb==0.6.0
click==8.1.7
cloudpickle==3.1.0
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.3.0
cryptography==44.0.1
ctranslate2==4.5.0
cycler==0.12.1
dataclasses-json==0.6.7
debugpy==1.8.8
decorator==5.1.1
deepgram-sdk==3.9.0
defusedxml==0.7.1
Deprecated==1.2.15
deprecation==2.1.0
distro==1.9.0
dm-tree==0.1.8
dnspython==2.7.0
durationpy==0.9
elevenlabs==1.51.0
environment==1.0.0
etils==1.5.2
exceptiongroup==1.2.2
executing==2.1.0
fastapi==0.115.6
faster-whisper==1.1.1
fastjsonschema==2.20.0
filelock==3.13.1
finance-datareader==0.9.94
Flask==3.0.3
flatbuffers==24.3.25
folium==0.19.0
fonttools==4.54.1
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2024.2.0
gast==0.6.0
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.37.0
google-crc32c==1.6.0
google-images-download==2.8.0
google-pasta==0.2.0
googleapis-common-protos==1.66.0
greenlet==3.1.1
grpcio==1.67.1
gym==0.26.2
gym-notices==0.0.8
h11==0.14.0
h5py==3.12.1
httpcore==1.0.7
httptools==0.6.4
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.27.0
humanfriendly==10.0
idna==3.10
ifaddr==0.2.0
imath==0.0.2
importlib_metadata==8.5.0
importlib_resources==6.4.5
ipykernel==6.29.5
ipython==8.18.1
ipython-genutils==0.2.0
ipywidgets==8.1.5
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.4
jiter==0.8.2
joblib==1.4.2
jpype1==1.5.1
json5==0.9.28
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-contrib-core==0.4.2
jupyter-contrib-nbextensions==0.7.0
jupyter-events==0.10.0
jupyter-highlight-selected-word==0.2.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_nbextensions_configurator==0.6.4
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
keras==3.6.0
kiwisolver==1.4.7
konlpy==0.6.0
kubernetes==31.0.0
langchain==0.3.18
langchain-community==0.3.14
langchain-core==0.3.35
langchain-text-splitters==0.3.6
langsmith==0.2.7
libclang==18.1.1
lxml==5.3.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.4.1
mmh3==5.0.1
monotonic==1.6
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
namex==0.0.8
narwhals==1.17.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.2.1
noisereduce==3.0.3
notebook==7.2.2
notebook_shim==0.2.4
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.19.2
openai==1.62.0
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
OpenEXR==3.3.2
opentelemetry-api==1.29.0
opentelemetry-exporter-otlp-proto-common==1.29.0
opentelemetry-exporter-otlp-proto-grpc==1.29.0
opentelemetry-instrumentation==0.50b0
opentelemetry-instrumentation-asgi==0.50b0
opentelemetry-instrumentation-fastapi==0.50b0
opentelemetry-proto==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-semantic-conventions==0.50b0
opentelemetry-util-http==0.50b0
opt_einsum==3.4.0
optree==0.13.1
orjson==3.10.13
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pillow==11.0.0
platformdirs==4.3.6
plotly==5.24.1
posthog==3.7.4
prometheus_client==0.21.0
promise==2.3
prompt_toolkit==3.0.48
propcache==0.2.1
protobuf==5.29.2
psutil==6.1.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAudio==0.2.14
pycparser==2.22
pydantic==2.10.3
pydantic-settings==2.7.1
pydantic_core==2.27.1
pydeck==0.9.1
pydub==0.25.1
pyee==12.1.1
Pygments==2.18.0
pylibsrtp==0.11.0
pyOpenSSL==25.0.0
pyparsing==3.2.0
PyPDF2==3.0.1
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.7
pytube==15.0.0
pytz==2024.2
pywin32==308
pywinpty==2.0.14
PyYAML==6.0.2
pyzmq==26.2.0
read==0.0.2
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.21.0
rsa==4.9
safetensors==0.4.5
scikeras==0.13.0
scikit-learn==1.5.2
scipy==1.13.1
seaborn==0.13.2
selenium==4.27.1
Send2Trash==1.8.3
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.5.1
soupsieve==2.6
SQLAlchemy==2.0.36
srt==3.5.3
stack-data==0.6.3
starlette==0.41.3
streamlit==1.41.0
streamlit-webrtc==0.47.9
streamlit_folium==0.23.2
sympy==1.13.1
tenacity==9.0.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-datasets==4.9.3
tensorflow-io-gcs-filesystem==0.31.0
tensorflow-metadata==1.16.1
tensorflow_intel==2.18.0
termcolor==2.5.0
terminado==0.18.1
threadpoolctl==3.5.0
tiktoken==0.8.0
tinycss2==1.4.0
tokenizers==0.21.0
toml==0.10.2
tomli==2.1.0
torch==2.5.1+cpu
torchaudio==2.5.1+cpu
torchvision==0.20.1+cpu
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.47.1
trio==0.27.0
trio-websocket==0.11.1
typer==0.15.1
types-python-dateutil==2.9.0.20241003
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
ultralytics==8.3.43
ultralytics-thop==2.0.12
uri-template==1.3.0
urllib3==2.2.3
uvicorn==0.34.0
vosk==0.3.45
watchdog==6.0.0
watchfiles==1.0.3
wavio==0.0.9
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==14.1
Werkzeug==3.1.2
widgetsnbextension==4.0.13
wordcloud==1.9.4
wrapt==1.16.0
wsproto==1.2.0
xyzservices==2024.9.0
yarl==1.18.3
zipp==3.21.0
